Big Commit. Not Finished.
I made lots of small changes. The biggest thing I did was at the start to an easy enemy creation system where you could get cookie cutter enemy movement types similar to Megaman and in a consistent scale without having to make lots of specifications. I would like to explain this more over the call, but I think its pretty cool and will ve super helpful. It is not complete, but it also does not have errors so you guys can try it and see what you think need to be changed.